pyproject.toml
[tool.poetry]
name = "cistercian-dates-nft"
version = "0.1.0"
description = ""
authors = ["Rasul Kireev"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.3.2"
eth-brownie = "^1.16.3"
python-dotenv = "0.16.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
